Add diag(), which implements Test::More's diag() taking a printf()-like format string argument.
Adjust _gen_result() to take three new parameters, corresponding to the
file, function, and line number where the test was called.  Use this 
information to report it using diag() if the test failed.

Remove the comment printed by _expected_tests(), it gets in the way of
actual testing.

Redefine ok() and ok2() to take advantage of the 3 new parameters to
_gen_result().


git-svn-id: svn://jc.ngo.org.uk/nik/libtap/trunk@198 2d5babbe-c3d5-0310-a406-eb661dfe2ac6