Implement ticket #4, and produce errors when running more (or fewer) tests than planned.
A number of changes in support of this.

tap_c:_tap_init(), make stdout unbuffered.

tap.c: _tap_init() call moved to each of the plan_*() functions so that it
runs as soon as possible.

tap.c:plan_tests(), remove a diagnostic message to be compatible with
Test::More, and only set have_plan when we're certain a valid plan has
been provided.

tap.c:exit_status(), return the correct exit status for more conditions
now, including when more or fewer tests than planned are run.

tap.c:_cleanup(), produce a diagnostic if a test dies before producing
any output.  Produce diagnostics if more or fewer tests than planned are
run.

README: Document that stdout is unbuffered after calling any of the plan_*()
functions.

Add tests that check that more tests than planned (too-many) and fewer
tests than planned (not-enough) were run.

Update test.t everywhere.  It now captures stderr as well as stdout for
comparison.  To avoid false positives for valid differences, output that
matches

    /^#     Failed test \(.*\)/

is normalised.  The text in the parentheses contains filenames and line
numbers which are legitimately different for the Perl and C test 
implementations.

Produce unified diff output when there are differences.


git-svn-id: svn://jc.ngo.org.uk/nik/libtap/trunk@214 2d5babbe-c3d5-0310-a406-eb661dfe2ac6