Update the README, note that pass() and fail() should not normally be used.
Update all the tests to check the return code from each function, and issue
it as a diagnostic message.  These are checked as part of the normal 'make
sure the output from the Perl and C implementations is identical' test,
so ensures that the behaviour is the same.  Closes #13.

Add tests for diag().  Closes #18.


git-svn-id: svn://jc.ngo.org.uk/nik/libtap/trunk@218 2d5babbe-c3d5-0310-a406-eb661dfe2ac6